+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## What's Changed
 
-* Add the handlers to the graph with `--show-handlers` (**the initial support for handlers**). They are by default added
+* feat: Add the handlers to the graph with `--show-handlers` (**the initial support for handlers**). They are by default added
   at the end of the play and roles only. This might change in the future to actually reflect the handlers' behavior.
 * Changes the shape of the graphviz node to make it consistent with Mermaid. The tasks will be rectangle instead of
   `octagon`: https://graphviz.org/doc/info/shapes.html
@@ -10,6 +10,9 @@
 * Add a new `display_name()` method to the node for a friendly name for the graph. This removes passing the
   `node_label_prefix` in multiple places.
 * Remove the play name from the edge going from playbook to the plays. This was not consistent with the other edges.
+* feat: Add a new option `--title` to add a title to the graph by @haidaraM
+  in https://github.com/haidaraM/ansible-playbook-grapher/pull/229. Default to 'Ansible Playbook Grapher'. The graphviz
+  renderer will now use this as the title (label). The Mermaid renderer already has a title.
 
 **Full Changelog**: https://github.com/haidaraM/ansible-playbook-grapher/compare/v2.7.0...v2.8.0